Updates watchlist to periodically refresh all entries
